run code inspection + can pick up liquids from magic source + fix dragon beacon not working + only show xp cost for abilities if the button is hovered (not when focused) + make only forest dragon immune to cactus damage + some utility methods for claw inventory + dont fetch capability for render newborn check + use entity type tag for animal avoid goal blacklist + run datagen

+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/blocks/DragonBeaconRenderer.java
@@ -23,7 +23,7 @@ public class DragonBeaconRenderer implements BlockEntityRenderer<DragonBeaconTil
 
 	@Override
 	public void render(DragonBeaconTileEntity dragonBeaconEntity, float v, PoseStack PoseStack, MultiBufferSource iRenderTypeBuffer, int light, int overlay){
-		dragonBeaconEntity.tick += 0.5;
+		dragonBeaconEntity.tick += 0.5f;
 		PoseStack.pushPose();
 		DragonBeaconTileEntity.Type type = dragonBeaconEntity.type;
 
@@ -61,7 +61,7 @@ public class DragonBeaconRenderer implements BlockEntityRenderer<DragonBeaconTil
 			}
 		}
 
-		float f1 = Mth.sin(((float)dragonBeaconEntity.tick + v) / 20.0F + dragonBeaconEntity.bobOffs) * 0.1F + 0.1F;
+		float f1 = Mth.sin((dragonBeaconEntity.tick + v) / 20.0F + dragonBeaconEntity.bobOffs) * 0.1F + 0.1F;
 		PoseStack.translate(0.5, 0.25 + f1 / 2f, 0.5);
 		PoseStack.mulPose(Axis.YP.rotationDegrees(dragonBeaconEntity.tick));
 		PoseStack.scale(2, 2, 2);
